refactored remaining types
