+++ src/server/api/routers/session.ts
@@ -4,18 +4,8 @@ import {
   publicProcedure,
   protectedProcedure,
 } from "~/server/api/trpc";
+import { generateUniqueCode } from "~/utils/ids";
 
-function generateUniqueCode(): string {
-  const characters = '0123456789';
-  let result = '';
-
-  for (let i = 0; i < 6; i++) {
-    const randomIndex = Math.floor(Math.random() * characters.length);
-    result += characters[randomIndex];
-  }
-
-  return result;
-}
 
 export const sessionRouter = createTRPCRouter({
   createSession: publicProcedure
@@ -33,7 +23,7 @@ export const sessionRouter = createTRPCRouter({
 
       let uniqueCode;
       do {
-        uniqueCode = generateUniqueCode();
+        uniqueCode = generateUniqueCode(6);
       } while (
         await ctx.prisma.createdSession.findUnique({
           where: {
@@ -68,7 +58,7 @@ export const sessionRouter = createTRPCRouter({
     .mutation(async ({ input, ctx }) => {
       const session = await ctx.prisma.createdSession.update({
         where: {
-          uniqueCode: Number(input.uniqueCode),
+          uniqueCode: input.uniqueCode,
         },
         data: {
           valid: false,
@@ -78,7 +68,7 @@ export const sessionRouter = createTRPCRouter({
       // End all associated examSessions
       await ctx.prisma.examSession.updateMany({
         where: {
-          uniqueCode: Number(input.uniqueCode),
+          uniqueCode: input.uniqueCode,
         },
         data: {
           endTime: new Date(), // setting endTime to the current date and time
@@ -94,7 +84,7 @@ export const sessionRouter = createTRPCRouter({
     .query(async ({ input, ctx }) => {
       const examSession = await ctx.prisma.examSession.findFirst({
         where: {
-          uniqueCode: parseInt(input.uniqueCode),
+          uniqueCode:input.uniqueCode,
           studentId: input.studentId,
         },
       });
@@ -114,7 +104,7 @@ export const sessionRouter = createTRPCRouter({
     .mutation(async ({ input, ctx }) => {
       const validSession = await ctx.prisma.createdSession.findUnique({
         where: {
-          uniqueCode: Number(input.uniqueCode),
+          uniqueCode: input.uniqueCode,
           valid: true,
         },
       });
